requirements.txt
fastapi
uvicorn
pillow
pandas
openpyxl
#pip freeze